+++ R/spatial.R
@@ -28,7 +28,7 @@ GetCentroids <- function(object, cells.use = NULL, get.exact = TRUE) {
   #Error checking
   cell.names <- intersect(x = cells.use, y = colnames(x = object@spatial@final.prob))
   if (length(x = cell.names) != length(x = cells.use)) {
-    print(paste(
+    message(paste(
       "Error",
       setdiff(x = cells.use, y = colnames(x = object@spatial@final.prob)),
       " have not been mapped"